questao 5 da lista

+++ lista01/ex03.c
@@ -8,7 +8,7 @@ int main()
  for (cont1 = 1,cont2 = 1; cont1 < final; cont1 += 2, cont2++)
  {
   printf("%.2f / %.2f \n", cont1,cont2);
-  result += (float)(cont1 / cont2);
+  result += (cont1 / cont2);
  }
   printf("= %.2f\n", result);
